[Feature] Fix TicketEntitie

+++ BackEnd/src/main/java/com/estacsis/Repository/TicketRepository.java
@@ -2,5 +2,5 @@ package com.estacsis.Repository;
 import com.estacsis.Entitie.TicketEntitie;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface TicketRepository extends JpaRepository<TicketEntitie, Integer> {
+public interface TicketRepository extends JpaRepository<TicketEntitie, Long> {
 }
